requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==23.1.0
click==8.1.3
coverage==7.2.3
decorator==5.1.1
Django==4.1.7
django-crispy-forms==2.0
django-extensions==3.2.1
djangorestframework==3.14.0
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
executing==1.2.0
inflection==0.5.1
ipython==8.12.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
postgis==1.0.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
uritemplate==4.1.1
wcwidth==0.2.6
